feat(orderreducer):connected to root, moved actions)
