+++ app/reducers/index.jsx
@@ -1,11 +1,13 @@
 import { combineReducers } from 'redux'
 import categoryReducer from './categories'
 import productReducer from './products'
+import orderReducer from './orders'
 
 const rootReducer = combineReducers({
   auth: require('./auth').default,
   categories: categoryReducer,
-  products: productReducer
+  products: productReducer,
+  orders: orderReducer
 })
 
 export default rootReducer
